+++ CHANGES.md
@@ -158,6 +158,15 @@ remnants of special-case VMS support have been removed.
 1. OpaqueMoveThreshold values >= 200 (the default) are now treated as
     infinite, and so will always cause opaque moving.
 
+1. Added various config parameters as inverses of existing params.  New
+   params (with existing param they invert in parens):
+    * `BackingStore` (`NoBackingStore`)
+    * `GrabServer` (`NoGrabServer`)
+    * `StartInButtonState` (`StartInMapState`)
+    * `NoSortIconManager` (`SortIconManager`)
+    * `NoRestartPreviousState` (`RestartPreviousState`)
+    * `NoDecorateTransients` (`DecorateTransients`)
+
 ### Internals
 
 1. A new code style has been chosen and the entire codebase reformatted